Split data into training and test set
- Split data into training and test set to allow for training using
optimization algorithms and separate testing. Data is split using leave
one out method, using a single database as a hold out set. This is not
as ideal as full group K fold cross validation, however that is not
feasible due to the amount of time taken for optimization on each pass