Re-implemented Group cross-validation in SFS
- Evaluation of SFS iterations are now performed using 3 fold group
cross validation.
- Data now initially split into training/validation and test sets using
on iteration of the LeavOneOut cross-validator iterator from sklearn.
This may need to be changed to adjust the balance between training and
test set sizes at a later time.
- Alterations do not need to be made to resampling method due to
train-validation-test splitting, as previously thought. This is due to
the grouped splitting of data, maintaining the balance between normal
and abnormal samples